Create Restart_SFXCodepackage.ps1
This is a fully automated script designed to restart services or programs running on production servers by validating logs such as performance counters or metrics. Additionally I've also implemented to send an E-Mail to the concerned team regarding the restarts can also log the output to any  log data source.